Implemented Swagger and first Customer save working
Need to add the Owner copy constructor and Admin copy constructor and replicate how service saves customer.

+++ Robal/src/main/java/ac/project/Robal/repositories/AccountRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ac.project.Robal.models.Account;
 
-public interface AccountRepository extends JpaRepository<Account,Integer>{
+
+public interface AccountRepository extends JpaRepository<Account,Long>{
 
 }
